use int64 for chat attachment sizes (#6657)
* use int64 for chat attachment sizes

+++ go/chat/server.go
@@ -1115,7 +1115,7 @@ func (h *Server) postAttachmentLocal(ctx context.Context, arg postAttachmentArg)
 		ctx = s3.NewFakeS3Context(ctx)
 	}
 	chatUI := h.getChatUI(arg.SessionID)
-	progress := func(bytesComplete, bytesTotal int) {
+	progress := func(bytesComplete, bytesTotal int64) {
 		parg := chat1.ChatAttachmentUploadProgressArg{
 			SessionID:     arg.SessionID,
 			BytesComplete: bytesComplete,
@@ -1249,7 +1249,7 @@ func (h *Server) postAttachmentLocalInOrder(ctx context.Context, arg postAttachm
 		ctx = s3.NewFakeS3Context(ctx)
 	}
 	chatUI := h.getChatUI(arg.SessionID)
-	progress := func(bytesComplete, bytesTotal int) {
+	progress := func(bytesComplete, bytesTotal int64) {
 		parg := chat1.ChatAttachmentUploadProgressArg{
 			SessionID:     arg.SessionID,
 			BytesComplete: bytesComplete,
@@ -1457,7 +1457,7 @@ func (h *Server) downloadAttachmentLocal(ctx context.Context, arg downloadAttach
 	var identBreaks []keybase1.TLFIdentifyFailure
 	ctx = Context(ctx, arg.IdentifyBehavior, &identBreaks, h.identNotifier)
 	chatUI := h.getChatUI(arg.SessionID)
-	progress := func(bytesComplete, bytesTotal int) {
+	progress := func(bytesComplete, bytesTotal int64) {
 		parg := chat1.ChatAttachmentDownloadProgressArg{
 			SessionID:     arg.SessionID,
 			BytesComplete: bytesComplete,
